Improvements-
In the First Usage Case, I have made the following improvements:
a) Added an `owner_ variable` that is immutable and set in the constructor.
b) The unnecessary `if` statement that checked whether signer was equal to `owner_ ` has been removed. Instead, the function now simply returns the result of the comparison between signer and `owner_`. This makes the code simpler and more readable, while still achieving the same result.

In the Second Usage Case, I have made the following improvements:
a) Remove unnecessary comments: The comments on the functions are already self-explanatory, so some of the comments can be removed to reduce the clutter and make the code more readable.
b) Add function visibility: Adding function visibility (public, external, internal, private) can improve code readability and reduce the possibility of errors.


- Indented the code so that it looks neat.
- Added relevant comments to the code for better understanding and knowledge.
- Added an improved definition of Signature in the Introduction.
- Corrected spelling mistakes, grammatical errors & add full forms to certain acronyms.
- Did paraphrasing.
- Added punctuation marks wherever, required.
- Added conclusion to make the tutorial complete.
- Added/removed adequate spacing between paragraphs to ensure neatness.